frigginlook@gmail.com
So I'm onto something that is true and being tried in my life.  If I am talking to someone i compliment them on the talents i like about that friend. I Leave out the struggles they  battle. When I'm  down we im in the perfect place to lift someone up. It's a momentum of kind qualities. I believe if I focus on the good then the bad will just fix its self.  Like ignoring my little brother and he stops.